Initial rc_files commit with bashrc files
This is still pretty ugly and needs some work, but it has a hand-
ful of nice aliases and functions. The prompt still needs work.
I like seeing jobs and the git repo in the prompt, but it needs
to be cleaned up and has a lot of room for improvement.